+++ NEWS.md
@@ -1,3 +1,4 @@
 # AutoScore 1.1.0
 
-* Initial CRAN submission.
+* Added a copy of a function from the now-archived `coxed` package to ensure compatibility.
+  The function is used as-is and credited appropriately.